requirements.txt
w3lib==1.22.0
mitmproxy==5.0.0
better_exceptions==0.2.2
PyMySQL==0.10.0
parsel==1.6.0
six==1.15.0
redis==2.10.6
DBUtils==1.3
PyYAML==5.3.1